Updates SystemStatus tests
SystemStatus tests were updated as these initially neglected that the HealthCheck and EnvironmentDump depencencies required an application (i.e. Flask) and a path for reporting on health status and providing an environment dump. Testing that the HealthCheck and EnvironmentDump objects have the application and path injected into then required other tests to be updated.