Add entry point for launching REST API
Aim is to use uWSGI with Nginx for serving Flask application.

A wsgi.py file has been added, which sets up all the required dependencies in the inversion of control container, and launches the Flask application.

So far, the /round/<round_number> and /season endpoints are working, but the /healthCheck and /environmentDump endpoints are not.

For the application to work, an instance of Flask needed to be passed into the FootballResultsServer constructor. Potentially something similar may be required for the health check and the environment dump. This also required updating tests in test_server.py.