Improve testing of FootballResultsServer
Flask-Testing was introduced to improve the testing of FootballResultsServer. An improvement seen was the ability to validate that the start method invoked the run method once on the Flask object instance returned to FootballResultsServer by the inversion of control container.

The aim was to also test out the /season and /round/<round_number> endpoints, but it appears that with the inversion of control container does not make this possible without launching an actual server session.