Deploy configurations for CircleCI 2.0
As the CircleCI configuration file is version 2, the section that existed for deployments was for CircleCI 1.0. End result was that although the build succeeded, no deployment to Heroku ever happened.

The configuration has now been updated to be CircleCI 2.0 compliant. The file `setup-heroku.sh` has also been added as the CircleCI configuration requires this for pulling down the source code from Heroku.