Environment dump endpoint
The environment dump endpoint has been updated to call git to pull out information such as most recent commit SHA, commit author, and applicable tags.

This also removes the dependency on needing to have a JSON file with similar information contained in it, as this file was not being deployed, resulting in the environment dump endpoint returning an Internal Server Error.