Use PyGithub in environment dump endpoint
The previous attempt to shell out to the git executable did not work in Heroku as it appears that the git executable does not exist on the system the application is running on.

Instead PyGithub is being used for retrieving the most recent commit and the matching tag, if one exists.

This is reliant on the environment variable GITHUB_TOKEN being present on the system the application is running from. If this environment variable is not present, no call will be made to GitHub. Instead commit and tag properties will be set to None (JSON null).